+++ x/oracle/genesis_test.go
@@ -15,15 +15,18 @@ import (
 func TestExportInitGenesis(t *testing.T) {
 	input := keeper.CreateTestInput(t)
 
-	input.OracleKeeper.SetFeederDelegation(input.Ctx, keeper.ValAddrs[0], keeper.Addrs[1])
-	input.OracleKeeper.SetExchangeRate(input.Ctx, "pair1:pair2", sdk.NewDec(123))
-	input.OracleKeeper.SetAggregateExchangeRatePrevote(input.Ctx, keeper.ValAddrs[0], types.NewAggregateExchangeRatePrevote(types.AggregateVoteHash{123}, keeper.ValAddrs[0], uint64(2)))
-	input.OracleKeeper.SetAggregateExchangeRateVote(input.Ctx, keeper.ValAddrs[0], types.NewAggregateExchangeRateVote(types.ExchangeRateTuples{{Pair: "foo", ExchangeRate: sdk.NewDec(123)}}, keeper.ValAddrs[0]))
-	input.OracleKeeper.SetPair(input.Ctx, "pair1:pair1")
-	input.OracleKeeper.SetPair(input.Ctx, "pair2:pair2")
-	input.OracleKeeper.SetMissCounter(input.Ctx, keeper.ValAddrs[0], 10)
-	input.OracleKeeper.SetPairReward(input.Ctx, &types.PairReward{
-		Pair: "pair1:pair2",
+	input.OracleKeeper.FeederDelegations.Insert(input.Ctx, keeper.ValAddrs[0], keeper.Addrs[1])
+	input.OracleKeeper.ExchangeRates.Insert(input.Ctx, "pair1:pair2", sdk.NewDec(123))
+	input.OracleKeeper.Prevotes.Insert(input.Ctx, keeper.ValAddrs[0], types.NewAggregateExchangeRatePrevote(types.AggregateVoteHash{123}, keeper.ValAddrs[0], uint64(2)))
+	input.OracleKeeper.Votes.Insert(input.Ctx, keeper.ValAddrs[0], types.NewAggregateExchangeRateVote(types.ExchangeRateTuples{{Pair: "foo", ExchangeRate: sdk.NewDec(123)}}, keeper.ValAddrs[0]))
+	input.OracleKeeper.Pairs.Insert(input.Ctx, "pair1:pair1")
+	input.OracleKeeper.Pairs.Insert(input.Ctx, "pair2:pair2")
+	input.OracleKeeper.MissCounters.Insert(input.Ctx, keeper.ValAddrs[0], 10)
+	input.OracleKeeper.PairRewards.Insert(input.Ctx, 0, types.PairReward{
+		Pair:        "pair1:pair2",
+		Id:          0,
+		VotePeriods: 100,
+		Coins:       sdk.NewCoins(sdk.NewInt64Coin("test", 1000)),
 	})
 	genesis := oracle.ExportGenesis(input.Ctx, input.OracleKeeper)