+++ x/oracle/simulation/decoder_test.go
@@ -40,14 +40,17 @@ func TestDecodeDistributionStore(t *testing.T) {
 
 	pair := "btc:usd"
 
+	erBytes, err := exchangeRate.Marshal()
+	require.NoError(t, err)
+
 	kvPairs := kv.Pairs{
 		Pairs: []kv.Pair{
-			{Key: types.ExchangeRateKey, Value: cdc.MustMarshal(&sdk.DecProto{Dec: exchangeRate})},
-			{Key: types.FeederDelegationKey, Value: feederAddr.Bytes()},
-			{Key: types.MissCounterKey, Value: cdc.MustMarshal(&gogotypes.UInt64Value{Value: missCounter})},
-			{Key: types.AggregateExchangeRatePrevoteKey, Value: cdc.MustMarshal(&aggregatePrevote)},
-			{Key: types.AggregateExchangeRateVoteKey, Value: cdc.MustMarshal(&aggregateVote)},
-			{Key: types.GetPairKey(pair), Value: []byte{}},
+			{Key: []byte{0x1, 0x2, 0x3, 0x4, 0x5}, Value: erBytes},
+			{Key: []byte{0x2, 0x3, 0x4, 0x5, 0x6}, Value: feederAddr.Bytes()},
+			{Key: []byte{0x3, 0x4, 0x5, 0x6, 0x7}, Value: cdc.MustMarshal(&gogotypes.UInt64Value{Value: missCounter})},
+			{Key: []byte{0x4, 0x3, 0x5, 0x7, 0x8}, Value: cdc.MustMarshal(&aggregatePrevote)},
+			{Key: []byte{0x5, 0x6, 0x7, 0x8, 0x9}, Value: cdc.MustMarshal(&aggregateVote)},
+			{Key: append([]byte{0x6}, append([]byte(pair), 0x0)...), Value: []byte{}},
 			{Key: []byte{0x99}, Value: []byte{0x99}},
 		},
 	}